// js/stat-evaluation-player/src/generated/PressureTeamStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LabeledFloatSums } from "./LabeledFloatSums.ts";

export type PressureTeamStats = { tracked_time: number, defensive_half_time: number, offensive_half_time: number, neutral_time: number, labeled_time?: LabeledFloatSums, };